Renamed pipeline to PixelPipeline
