+++ Core/Core.vcxproj
@@ -28,7 +28,6 @@
     <ClInclude Include="Include\Joypad.h" />
     <ClInclude Include="Include\HighTimer.h" />
     <ClInclude Include="Include\Pch.h" />
-    <ClInclude Include="Include\Pipeline.h" />
     <ClInclude Include="Include\PixelPipeline.h" />
     <ClInclude Include="Include\PixelProcessor.h" />
     <ClInclude Include="Include\Ram.h" />
@@ -36,7 +35,6 @@
     <ClInclude Include="Include\Dma.h" />
   </ItemGroup>
   <ItemGroup>
-    <ClCompile Include="Source\Pipeline.cpp" />
     <ClCompile Include="Source\PixelPipeline.cpp" />
     <ClCompile Include="Source\PixelProcessor.cpp" />
     <ClCompile Include="Source\Joypad.cpp" />
