+++ RetroGBm/Shaders/PixelShader.hlsl.h
@@ -50,10 +50,10 @@ ret
 
 const BYTE g_PixelShader[] =
 {
-     68,  88,  66,  67, 179, 117, 
-     94, 213, 226, 135,  56, 116, 
-    172, 103,  10,  81,   5,  75, 
-    243, 190,   1,   0,   0,   0, 
+     68,  88,  66,  67, 138, 134, 
+    164, 198,   9,  49, 173, 113, 
+    169,   4, 110, 183,  62, 234, 
+     54,  34,   1,   0,   0,   0, 
     192,  56,   0,   0,   6,   0, 
       0,   0,  56,   0,   0,   0, 
       0,   1,   0,   0,  84,   1, 
@@ -594,11 +594,11 @@ const BYTE g_PixelShader[] =
       0,   0,   0,   0,   0,   0, 
       0,   0,   0,   0,   0,   0, 
       0,   0,   0,   0,   0,   0, 
-    148,  46,  49,   1,  36,  64, 
+    148,  46,  49,   1,  64, 104, 
     247, 101,   1,   0,   0,   0, 
-    140,  40, 114, 100, 210, 227, 
-    102,  64, 188,  96, 100,  19, 
-    209,  74,  85, 180,   0,   0, 
+     23,  77,  93, 218, 238, 114, 
+     28,  72, 166,  39, 191,  53, 
+    185,  64, 176,  29,   0,   0, 
       0,   0,   0,   0,   0,   0, 
       1,   0,   0,   0,   1,   0, 
       0,   0,   0,   0,   0,   0, 
@@ -769,9 +769,9 @@ const BYTE g_PixelShader[] =
      53,   5,   1,   0, 101, 210, 
       2,   0,  38, 247,   2,   0, 
      65,  36,   1,   0,  39,  82, 
-      3,   0,  34,  80,   1,   0, 
+      3,   0, 130, 182,   0,   0, 
     217,  42,   2,   0,  21, 230, 
-      0,   0,  23, 114,   2,   0, 
+      0,   0,  20, 210,   2,   0, 
      73,  20,   1,   0, 153, 189, 
       3,   0,   0,  16,   0,   0, 
       0,   0,   0,   0,   0,   0, 
@@ -1107,8 +1107,8 @@ const BYTE g_PixelShader[] =
     115, 105, 116, 105, 111, 110, 
      32,  58,  32,  80,  79,  83, 
      27, 226,  48,   1, 172,   0, 
-      0,   0, 239, 157, 247, 197, 
-    158, 120, 218,   1,   1,   0, 
+      0,   0, 103,  57, 119, 174, 
+    182, 120, 218,   1,   1,   0, 
       0,   0,   0,   0,   0,   0, 
       0,   0,   0,   0,   0,   0, 
       0,   0,   0,   0,   0,   0, 
@@ -1400,14 +1400,14 @@ const BYTE g_PixelShader[] =
       0,   0,  23,   0,   1,   0, 
       4,  16,   0,   0,  14,   0, 
      23,  21,   0,   0,   0,   0, 
-     10,   2, 160, 164,   0,   0, 
+     10,   2,  64, 108,   0,   0, 
     242, 241,  10,   0,  24,  21, 
       7,  16,   0,   0,   1,   0, 
       1,   0,  10,   0,  24,  21, 
       8,  16,   0,   0,   1,   0, 
       0,   2,  14,   0,  23,  21, 
       0,  16,   0,   0,   3,   2, 
-    144,  83,   0,   0, 242, 241, 
+    176,  64,   0,   0, 242, 241, 
      10,   0,  24,  21,  10,  16, 
       0,   0,   1,   0,   1,   0, 
      10,   0,  24,  21,  11,  16, 
@@ -2045,11 +2045,11 @@ const BYTE g_PixelShader[] =
       0,   0,   0,   0,   0,   0, 
       0,   0,   0,   0,   0,   0, 
       0,   0,   0,   0, 148,  46, 
-     49,   1,  36,  64, 247, 101, 
-      1,   0,   0,   0, 140,  40, 
-    114, 100, 210, 227, 102,  64, 
-    188,  96, 100,  19, 209,  74, 
-     85, 180, 164,   0,   0,   0, 
+     49,   1,  64, 104, 247, 101, 
+      1,   0,   0,   0,  23,  77, 
+     93, 218, 238, 114,  28,  72, 
+    166,  39, 191,  53, 185,  64, 
+    176,  29, 164,   0,   0,   0, 
      47,  76, 105, 110, 107,  73, 
     110, 102, 111,   0,  47, 110, 
      97, 109, 101, 115,   0,  47,
