+++ RetroGBm/Shaders/VertexShader.hlsl.h
@@ -49,10 +49,10 @@ ret
 
 const BYTE g_VertexShader[] =
 {
-     68,  88,  66,  67,  21,  48, 
-     95, 206, 130,  79, 149, 247, 
-    204, 124, 108, 173,  25,   4, 
-    203, 140,   1,   0,   0,   0, 
+     68,  88,  66,  67, 118,   7, 
+    193, 107, 139,  47, 143, 161, 
+     55, 237, 153,  50, 146, 107, 
+     52,  86,   1,   0,   0,   0, 
     168,  56,   0,   0,   6,   0, 
       0,   0,  56,   0,   0,   0, 
     164,   0,   0,   0, 248,   0, 
@@ -589,11 +589,11 @@ const BYTE g_VertexShader[] =
       0,   0,   0,   0,   0,   0, 
       0,   0,   0,   0,   0,   0, 
       0,   0,   0,   0,   0,   0, 
-    148,  46,  49,   1,  36,  64, 
+    148,  46,  49,   1,  65, 104, 
     247, 101,   1,   0,   0,   0, 
-    106,  70, 224,  49,  31, 254, 
-    122,  67, 136, 148, 215,  47, 
-     65, 221, 185, 126,   0,   0, 
+    180,  18,  16, 102,  50,  82, 
+    223,  76, 150,  51, 226, 145, 
+    238, 183, 212, 154,   0,   0, 
       0,   0,   0,   0,   0,   0, 
       1,   0,   0,   0,   1,   0, 
       0,   0,   0,   0,   0,   0, 
@@ -1102,8 +1102,8 @@ const BYTE g_VertexShader[] =
     116,  59,  13,  10, 125,   0, 
      47,  47,  32,  86, 101, 114, 
      27, 226,  48,   1, 172,   0, 
-      0,   0, 121, 156,  29, 198, 
-    158, 120, 218,   1,   1,   0, 
+      0,   0, 132,  17, 166, 174, 
+    182, 120, 218,   1,   1,   0, 
       0,   0,   0,   0,   0,   0, 
       0,   0,   0,   0,   0,   0, 
       0,   0,   0,   0,   0,   0, 
@@ -2125,11 +2125,11 @@ const BYTE g_VertexShader[] =
       0,   0,   0,   0,   0,   0, 
       0,   0,   0,   0,   0,   0, 
       0,   0,   0,   0,   0,   0, 
-    148,  46,  49,   1,  36,  64, 
+    148,  46,  49,   1,  65, 104, 
     247, 101,   1,   0,   0,   0, 
-    106,  70, 224,  49,  31, 254, 
-    122,  67, 136, 148, 215,  47, 
-     65, 221, 185, 126, 165,   0, 
+    180,  18,  16, 102,  50,  82, 
+    223,  76, 150,  51, 226, 145, 
+    238, 183, 212, 154, 165,   0, 
       0,   0,  47,  76, 105, 110, 
     107,  73, 110, 102, 111,   0, 
      47, 110,  97, 109, 101, 115,
